backend/requirements.txt
flask
flask-cors
anthropic
yfinance
pandas
numpy
python-dotenv
requests